backend/requirements.txt
django==4.2.13
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2

